spatPatClassifyR: Classify Spatial Landscape Patterns Using Neural Networks

Classification of spatial landscape patterns using neural networks. The package provides tools for generating artificial landscapes with different spatial patterns, calculating landscape metrics, and training neural network classifiers.

Two classification approaches are supported:

**Pixel-based classification** using convolutional neural networks via keras3 (see train_nn_pixels)

**Metrics-based classification** using landscape metrics computed with landscapemetrics as input features for a neural network (see train_nn_metrics)

Typical workflow

Generate training landscapes with create_landscapes

Optionally calculate and evaluate landscape metrics with calculate_landscape_metrics and evaluate_landscape_metrics

Train a classifier with train_nn_pixels or train_nn_metrics

Apply the trained model to new landscapes with apply_nn_pixels or apply_nn_metrics

Visualize results with plot_classified_landscapes

References

Baldauf, S., Tietjen, B., & Berger, U. (2025). spatPatClassifyR: An R package for classifying spatial landscape patterns using neural networks. *Methods in Ecology and Evolution*. In review.

See also

Author

Maintainer: Selina Baldauf selina.baldauf@fu-berlin.de (ORCID)

Authors:

Britta Tietjen britta.tietjen@fu-berlin.de (ORCID)

Uta Berger uta.berger@tu-dresden.de (ORCID)